fastapi_be/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiosqlite==0.17.0; python_version >= '3.6'
anyio==3.3.4; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
asyncpg==0.24.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3; python_version >= '3.6'
fastapi==0.70.0
h11==0.12.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
iso8601==0.1.16
passlib==1.7.4
psycopg2-binary==2.9.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.8.2; python_full_version >= '3.6.1'
pyjwt==2.3.0
pypika-tortoise==0.1.1; python_version >= '3.7' and python_version < '4.0'
python-multipart==0.0.5
pytz==2021.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.16.0; python_version >= '3.6'
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
